[bugfix]: Don't use 0 as the default chainId && make chainId one u64 field (#83)

+++ dependencies/aptos-types/src/account_config/resources/chain_id.rs
@@ -12,7 +12,7 @@ use serde::Deserialize;
 
 #[derive(Deserialize)]
 pub struct ChainIdResource {
-    chain_id: u8,
+    chain_id: u64,
 }
 
 impl ChainIdResource {
